deployment/requirements.txt
a2a-sdk>=0.3.0
google-adk>=1.0.0
google-genai>=1.0.0
pydantic>=2.0.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
